Analytics Behaviour
This adds a Healthlocker.Analytics behaviour with two methods:
`identify` and `track`. There is one implementation of this behaviour:
`Local`. This local implementation will just log out the information,
this is for development and test environments.

There is also an EventNameBuilder and PropertiesBuilder to allow us to
flexibly add new events and properties to be tracked.