Render room and message components server side
When a user first visits a room we were originally sending JSON for
the messages to the client, which would then render the template.
However, as we now have specific components (which have some logic),
we'd rather render everything server-side for consistency.

This changes the RoomChannel to render the template, and broadcast that
to all clients. There's some complexity on the client to add the
relevant classes to a message, depending upon whether the client is the
sender or the receiver.