Remove markdown from pages where it relies on user input #1074
Markdown converts html & markdown to html, so *any* html that is input will be interpreted as such. These sections don't need markdown because they are mostly single sentence inputs that don't require additional formatting, and there is also not user interface for how to use markdown even when it is useful for formatting.